# docs/swagger.yaml
basePath: /api
definitions:
  web.ProjectCreateRequest:
    properties:
      confidence:
        maximum: 100
        minimum: 0
        type: number
      description:
        type: string
      name:
        type: string
      progress:
        maximum: 100
        minimum: 0
        type: number
      trend:
        enum:
        - up
        - down
        - stable
        type: string
      user_id:
        type: string
    required:
    - name
    - user_id
    type: object
  web.ProjectResponse:
    properties:
      confidence:
        type: number
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      progress:
        type: number
      trend:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  web.ProjectUpdateRequest:
    properties:
      confidence:
        maximum: 100
        minimum: 0
        type: number
      description:
        type: string
      id:
        type: string
      name:
        type: string
      progress:
        maximum: 100
        minimum: 0
        type: number
      trend:
        enum:
        - up
        - down
        - stable
        type: string
    required:
    - id
    - name
    type: object
  web.TaskCreateRequest:
    properties:
      bottleneck:
        type: string
      deliverable:
        type: string
      difficulty_level:
        type: string
      effort:
        type: integer
      priority:
        enum:
        - low
        - medium
        - high
        type: string
      project_id:
        type: string
      status:
        enum:
        - todo
        - in-progress
        - completed
        type: string
      title:
        type: string
    required:
    - effort
    - project_id
    - title
    type: object
  web.TaskResponse:
    properties:
      bottleneck:
        type: string
      created_at:
        type: string
      deliverable:
        type: string
      difficulty_level:
        type: string
      effort:
        type: integer
      id:
        type: string
      priority:
        type: string
      project_id:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  web.TaskUpdateRequest:
    properties:
      bottleneck:
        type: string
      deliverable:
        type: string
      difficulty_level:
        type: string
      effort:
        type: integer
      priority:
        enum:
        - low
        - medium
        - high
        type: string
      status:
        enum:
        - todo
        - in-progress
        - completed
        type: string
      title:
        type: string
    required:
    - effort
    - priority
    - status
    - title
    type: object
...
  web.WebResponse:
    properties:
      code:
        type: integer
      data: {}
      status:
        type: string
    type: object
host: localhost:3001
info:
  contact: {}
  description: API untuk manajemen users dengan PostgreSQL.
  title: User API
  version: "1.0"
paths:
...
  /projects:
    get:
      consumes:
      - application/json
      description: Get all projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/web.ProjectResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get all projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Create new project with the input payload
      parameters:
      - description: Create project request
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/web.ProjectCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.ProjectResponse'
      security:
      - BearerAuth: []
      summary: Create new project
      tags:
      - projects
  /projects/{id}:
    delete:
      consumes:
      - application/json
      description: Delete project by ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: response with code and status
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - projects
    get:
      consumes:
      - application/json
      description: Get project by ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.ProjectResponse'
      security:
      - BearerAuth: []
      summary: Get project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Update project by ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Update project request
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/web.ProjectUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.ProjectResponse'
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - projects
  /projects/user/{userId}:
    get:
      consumes:
      - application/json
      description: Get all projects for a specific user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/web.ProjectResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get projects by user ID
      tags:
      - projects
...
  /tasks:
    get:
      consumes:
      - application/json
      description: Get all tasks in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/web.TaskResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get all tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task with the provided data
      parameters:
      - description: Task data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/web.TaskCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.TaskResponse'
      security:
      - BearerAuth: []
      summary: Create a new task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: Delete a task
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: Get a task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.TaskResponse'
      security:
      - BearerAuth: []
      summary: Get a task by ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Update a task with the provided data
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Task data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/web.TaskUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.TaskResponse'
      security:
      - BearerAuth: []
      summary: Update a task
      tags:
      - tasks
  /tasks/project/{projectId}:
    get:
      consumes:
      - application/json
      description: Get all tasks for a specific project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/web.TaskResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get tasks by project ID
      tags:
      - tasks
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"